test: 잔고 테스트를 order.test.js에서 balance.test.js로 분리 (#149)
order.test.js에 쌓인 잔고 조회 테스트(buildBalanceParams, fetchAllBalance,
formatBalanceReport, formatTruncationNote, formatPercent, 잔고 파서 오염 방지)를
새 balance.test.js로 순수 이동한다. order.test.js에는 주문 관련(buildCashOrderBody,
resolveStock, 주문 가드)만 남는다. 단언 변경·추가 없음 — 이동 전 43건 = 이동 후
20(order) + 23(balance). node --test 43 pass.
